pet-in-fastapi/requirements.txt
fastapi==0.115.4
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
uvicorn==0.32.0
python-dotenv==1.0.1